+++ src/main/java/roomescape/application/dto/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto.request;
+package roomescape.application.dto;
 
 import java.time.LocalTime;
 import roomescape.domain.ReservationTime;
