+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.application.dto.request.ReservationTimeRequest;
-import roomescape.application.dto.response.ReservationTimeResponse;
+import roomescape.application.dto.ReservationTimeRequest;
+import roomescape.application.dto.ReservationTimeResponse;
 import roomescape.application.service.ReservationTimeService;
 
 @RestController
