+++ src/test/java/roomescape/application/service/ReservationTimeServiceTest.java
@@ -1,0 +1,5 @@
+package roomescape.application.service;
+
+class ReservationTimeServiceTest {
+
+}
