+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -15,7 +15,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.application.dto.response.ReservationResponse;
+import roomescape.application.dto.ReservationResponse;
 import roomescape.domain.ReservationTime;
 
 @SpringBootTest(
